# docs/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
...
    TrainerInfoByIdResponse:
      type: object
      required:
      - id
      - title
      - is_active
      - created_at
      - updated_at
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        difficulty_level:
          type: string
        is_active:
          type: boolean
        created_at:
          $ref: "#/components/schemas/LocalDateTime"
        updated_at:
          $ref: "#/components/schemas/LocalDateTime"
...
paths:
...
  /trainers/{id}:
    get:
      summary: Получение информации о тренажере по ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainerInfoByIdResponse"
      tags:
      - Trainer Controller
...
  /trainers/{id}/reset:
    post:
      summary: Сброс прогресса пользователя по ID тренажера
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204": {}
      tags:
      - Trainer Controller
...
info:
  title: Analytics-trainer API
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value